Changed QualtyValueStrategy implemetations from 1 abstract class, 2 subclasses to 1 enum with 2 values.  fixed bug in DefaultContig.Builder
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@365 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ test/org/jcvi/glyph/qualClass/TestQualityClassBuilder.java
@@ -393,7 +393,7 @@ public class TestQualityClassBuilder {
         //1 high quality conflict
         sut.addHighQualityConflict(SequenceDirection.REVERSE);
         
-        assertCorrectQualityClassBuilt(QualityClass.valueOf((byte)15), sut.build());
+        assertCorrectQualityClassBuilt(QualityClass.valueOf(15), sut.build());
         
     }
     
@@ -406,6 +406,6 @@ public class TestQualityClassBuilder {
         .addHighQualityConflict(SequenceDirection.FORWARD)
         .addLowQualityConflict(SequenceDirection.FORWARD)
         .addLowQualityConflict(SequenceDirection.FORWARD);
-        assertCorrectQualityClassBuilt(QualityClass.valueOf((byte)17), sut.build());
+        assertCorrectQualityClassBuilt(QualityClass.valueOf(17), sut.build());
     }
 }
